textWidth fixed in all Button2D classes along the examples

+++ examples/Timing/CameraInterpolation/CameraInterpolation.pde
@@ -35,7 +35,7 @@ import remixlab.bias.event.*;
 
 Scene scene;
 ArrayList buttons;
-int h;
+float h;
 PFont buttonFont;
 
 void setup() {
